Add Arduino fan controller.
Controls computer PWM fans using an Arduino and the ROG_EXT module.

An Arduino Mega is used here, but can be reconfigured to other boards.

The fan power is taken from the motherboards HAMP connector, which supplies high current, so multiple fans can be controlled from that 1 port, and one of the fans tachometer line is connected to the motherboards HAMP connector so RPM monitoring is still possible.

The ROG_EXT module supplies the CPU temperature to the Arduino and the fans are turned on or off using a NPN transistor (2SD882 for example) to switch ground on the fans, and when the fans are turned on, the PWM is controlled based on the CPU temperature. The PWM is ramped up and down gradually to avoid unwanted fan noise. The fans will not turn back on for a few seconds after turning off, to also prevent unwanted noise.

The Arduino USB is connected to a motherboard USB2 header, making serial monitoring possible.